Move Action Cable configuration to top level
